implement loopWriteLocations in terms of map for better reasoning
NOTE: I really should let the Haskell flow through me when I use Coq,
because FP is so so so much easier to reason with in Coq, because they
have map/filter/fold in the stdlib with a bunch of stuff already proven
with them. This is an important lesson I should not forget.